pyproject.toml
[tool.poetry]
name = "minsim-reddit-crawler"
version = "0.1.0"
description = ""
authors = ["nohkwanok <cloq@kakao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
praw = "^7.7.1"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
redis = "^5.0.1"
setuptools = "^68.2.2"
aiohttp = "^3.8.6"
asyncpraw = "^7.7.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
ruff = "^0.0.292"
mypy = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ["*_pb2.py", "*_pb2_grpc.py", "*_pb2.pyi"]